pyproject.toml
[project]
name = "vague"
version = "0.1.0"
requires-python = ">=3.11"
description = "Python CLI layer for LLM skill-based AI workflows"

dependencies = [
    "typer>=0.12",
    "python-frontmatter>=1.1",
    "pyyaml>=6.0",
    "gitpython>=3.1",
    "pydantic>=2.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
vague-install = "vague.installer:install_app"
vague         = "vague.sdk.cli:sdk_app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["vague/assets/**/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
